Replace existing http json grafana dashboard with grafana http and ws dashboard, arbitrarily update vus per pod constant bc couldn’t access the appropriate special instance type nodes, update installation checks to use catch to return desired error message, move delete existing aws lbc policy role to edamame teardown process instead of edamame init, add back babel-plugin-transform-import-meta installation to package.json, so that jest tests can run, update the parallelism calculation in the jest test b/c the arguments passed to parallelism() function in utilities/manifest.js have changed, update outdated parts of readme

+++ src/utilities/eksctl.js
@@ -15,13 +15,11 @@ const eksctl = {
   existsOrError() {
     return (
       exec(`eksctl version`)
-        .then(({stdout}) => {
-          if (!stdout) {
-            const msg = `Eksctl isn't installed. Please install eksctl; ` +
-              `instructions can be found at: ` +
-              `https://docs.aws.amazon.com/eks/latest/userguide/eksctl.html`;
-            throw new Error(msg);
-          }
+        .catch(() => {
+          const msg = `Eksctl isn't installed. Please install eksctl; ` +
+            `instructions can be found at: ` +
+            `https://docs.aws.amazon.com/eks/latest/userguide/eksctl.html`;
+          throw new Error(msg);
         })
     );
   },
@@ -97,9 +95,11 @@ const eksctl = {
   },
 
   deleteOldIamLBCPolicy(policy) {
-    return exec(
-      `aws iam delete-policy --policy-arn ${policy}`
-    );
+    if (policy) {
+      return exec(
+        `aws iam delete-policy --policy-arn ${policy}`
+      );
+    }
   },
 
   createIamRoleLBCPol(policyArn) {
